# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio centralizado de autenticación y autorización con JWT RS256, RBAC y auditoría.",
        "title": "Sentinel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Retorna la lista paginada de aplicaciones cliente registradas en el sistema.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Listar aplicaciones",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Número de página (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elementos por página (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda por nombre o slug",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar por estado activo",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de aplicaciones",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerPaginatedApplications"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Registra una nueva aplicación cliente en el sistema y genera su clave secreta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Crear aplicación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos de la nueva aplicación",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerCreateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Aplicación creada exitosamente",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerApplicationItem"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o slug con formato incorrecto",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Ya existe una aplicación con este slug",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Retorna los detalles de una aplicación, incluyendo su clave secreta.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Obtener aplicación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la aplicación (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la aplicación",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerApplicationItem"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Aplicación no encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Actualiza el nombre y/o estado de una aplicación. La aplicación del sistema no puede ser modificada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Actualizar aplicación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la aplicación (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerUpdateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aplicación actualizada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerApplicationItem"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso o aplicación de sistema",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Aplicación no encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/applications/{id}/rotate-key": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Genera una nueva clave secreta para la aplicación, invalidando la anterior. La aplicación del sistema no puede rotar su clave por esta vía.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Rotar clave de aplicación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la aplicación (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nueva clave generada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerRotateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso o aplicación de sistema",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Aplicación no encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/cost-centers/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Actualiza el nombre y/o estado de un centro de costo existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Centros de Costo"
                ],
                "summary": "Actualizar centro de costo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del centro de costo (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerUpdateCostCenterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centro de costo actualizado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerCostCenterItem"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Sin permiso",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Centro de costo no encontrado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.SwaggerApplicationItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_system": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Mi Aplicación"
                },
                "slug": {
                    "type": "string",
                    "example": "mi-aplicacion"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:30:00Z"
                }
            }
        },
...
        "internal_handler.SwaggerCreateApplicationRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Mi Aplicación"
                },
                "slug": {
                    "type": "string",
                    "example": "mi-aplicacion"
                }
            }
        },
...
        "internal_handler.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_handler.SwaggerErrorDetail"
                }
            }
        },
...
        "internal_handler.SwaggerPaginatedApplications": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler.SwaggerApplicationItem"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 3
                },
                "total_pages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppKeyAuth": {
            "description": "Clave secreta de la aplicación cliente",
            "type": "apiKey",
            "name": "X-App-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Token JWT de acceso. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}